+++ email/src/watch/config.rs
@@ -1,5 +1,8 @@
 use process::Cmd;
 use serde::{Deserialize, Serialize};
+use std::{fmt, ops::Deref, sync::Arc};
+
+use crate::{envelope::Envelope, Result};
 
 /// Watch hook configuration.
 ///
@@ -17,6 +20,52 @@ pub enum WatchHook {
     /// Send a system notification using the given
     /// [`notify_rust::Notification`]-like configuration.
     Notify(WatchNotifyConfig),
+
+    /// Execute the given watch function.
+    ///
+    /// The watch function cannot be de/serialized. The function
+    /// should take a reference to an envelope and return a [`Result`]
+    /// of unit.
+    #[serde(skip)]
+    Fn(WatchFn),
+}
+
+/// Watch function.
+///
+/// This is just a wrapper around a function that takes a reference to
+/// an envelope.
+#[derive(Clone)]
+pub struct WatchFn(Arc<dyn Fn(&Envelope) -> Result<()> + Send + Sync>);
+
+impl WatchFn {
+    /// Create a new watch function.
+    pub fn new(f: impl Fn(&Envelope) -> Result<()> + Send + Sync + 'static) -> Self {
+        Self(Arc::new(f))
+    }
+}
+
+impl Deref for WatchFn {
+    type Target = Arc<dyn Fn(&Envelope) -> Result<()> + Send + Sync>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl fmt::Debug for WatchFn {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "WatchHookFn()")
+    }
+}
+
+impl Eq for WatchFn {
+    //
+}
+
+impl PartialEq for WatchFn {
+    fn eq(&self, _other: &Self) -> bool {
+        true
+    }
 }
 
 /// The watch configuration of the notify hook variant.