+++ email/tests/sendmail.rs
@@ -25,9 +25,7 @@ async fn test_sendmail_features() {
         port: 3143,
         encryption: Some(ImapEncryptionKind::None),
         login: "bob@localhost".into(),
-        auth: ImapAuthConfig::Passwd(PasswdConfig {
-            passwd: Secret::new_raw("password"),
-        }),
+        auth: ImapAuthConfig::Passwd(PasswdConfig(Secret::new_raw("password"))),
         ..Default::default()
     };
     let sendmail_config = SendmailConfig {